chore(*): Refactor a bit to fix tests.
