+++ projects/igniteui-angular/src/lib/grids/pivot-grid/pivot-util.ts
@@ -2,7 +2,8 @@ import { cloneValue } from '../../core/utils';
 import { DataUtil } from '../../data-operations/data-util';
 import { FilteringLogic } from '../../data-operations/filtering-expression.interface';
 import { FilteringExpressionsTree } from '../../data-operations/filtering-expressions-tree';
-import { ISortingExpression } from '../../data-operations/sorting-strategy';
+import { DefaultPivotSortingStrategy } from '../../data-operations/pivot-strategy';
+import { ISortingExpression, SortingDirection } from '../../data-operations/sorting-strategy';
 import { IGridSortingStrategy, IgxSorting } from '../common/strategy';
 import { IPivotConfiguration, IPivotDimension, IPivotGridRecord, IPivotKeys, IPivotValue, PivotDimensionType } from './pivot-grid.interface';
 
@@ -398,5 +399,25 @@ export class PivotUtil {
                 }
             }
         }
+    }
+
+    public static generateDimensionSortingExpressions(dimensions: IPivotDimension[]): ISortingExpression[] {
+        const expressions: ISortingExpression[] = [];
+        PivotUtil.flatten(dimensions).forEach(x => {
+            if (x.sortDirection) {
+                expressions.push({
+                    dir: x.sortDirection,
+                    fieldName: x.memberName,
+                    strategy: DefaultPivotSortingStrategy.instance()
+                });
+            } else {
+                expressions.push({
+                    dir: SortingDirection.None,
+                    fieldName: x.memberName,
+                    strategy: DefaultPivotSortingStrategy.instance()
+                });
+            }
+        });
+        return expressions;
     }
 }
